requirements.txt
finnhub-python
python-dotenv
pandas
openpyxl